% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bivden.R, R/print.msden.R, R/print.rrs.R,
%   R/print.rrst.R, R/print.stden.R
\name{print.bivden}
\alias{print.bivden}
\alias{print.rrs}
\alias{print.msden}
\alias{print.stden}
\alias{print.rrst}
\title{Printing sparr objects}
\usage{
\method{print}{bivden}(x, ...)

\method{print}{msden}(x, ...)

\method{print}{rrs}(x, ...)

\method{print}{rrst}(x, ...)

\method{print}{stden}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{bivden}}, \code{\link{stden}},
\code{\link{rrs}}, \code{\link{rrst}}, or \code{\link{msden}}.}

\item{...}{Ignored.}
}
\description{
\code{print} methods for classes \code{\link{bivden}}, \code{\link{stden}},
\code{\link{rrs}}, \code{\link{rrst}} and \code{\link{msden}}.
}
\author{
T.M. Davies
}
