\name{summary.bivden}
\Rdversion{1.1}
\alias{summary.bivden}
\alias{print.bivden}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarising a bivariate kernel density estimate object
}
\description{
\code{print} and \code{summary} methods for class \code{"bivden"}
}
\usage{
\method{print}{bivden}(x, \dots)
\method{summary}{bivden}(object, \dots)
}
\arguments{
  \item{x, object}{
An object of class \code{"bivden"}.
}
  \item{...}{
Ignored.
}
}

  
%\details{

%}
%\value{

%}
%\references{
%}
\author{
T.M. Davies
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
%\examples{
%KBivN(c(0.1,0.4))
%}

