% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.msden.R
\name{summary.msden}
\alias{summary.msden}
\alias{print.msden}
\title{Summarising an estimated multi-scale density object}
\usage{
\method{summary}{msden}(object, ...)
}
\arguments{
\item{...}{Ignored.}

\item{x, object}{An object of class \code{\link{msden}} resulting from a call
to \code{\link{multiscale.density}}.}
}
\description{
\code{print} and \code{summary} methods for class \code{\link{msden}}
}
\author{
T.M. Davies
}
