% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rrs.R
\name{summary.rrs}
\alias{summary.rrs}
\alias{print.rrs}
\title{Summarising an estimated relative risk function object}
\usage{
\method{summary}{rrs}(object, ...)
}
\arguments{
\item{...}{Ignored.}

\item{x, object}{An object of class \code{\link{rrs}} resulting from a call
to \code{\link{risk}}.}
}
\description{
\code{print} and \code{summary} methods for class \code{\link{rrs}}.
}
\author{
T.M. Davies
}
