% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bivden.R, R/summary.msden.R,
%   R/summary.rrs.R, R/summary.rrst.R, R/summary.stden.R
\name{summary.bivden}
\alias{summary.bivden}
\alias{summary.rrs}
\alias{summary.msden}
\alias{summary.stden}
\alias{summary.rrst}
\title{Summarising sparr objects}
\usage{
\method{summary}{bivden}(object, ...)

\method{summary}{msden}(object, ...)

\method{summary}{rrs}(object, ...)

\method{summary}{rrst}(object, ...)

\method{summary}{stden}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{bivden}}, \code{\link{stden}},
\code{\link{rrs}}, \code{\link{rrst}}, or \code{\link{msden}}.}

\item{...}{Ignored.}
}
\description{
\code{summary} methods for classes \code{\link{bivden}}, \code{\link{stden}},
\code{\link{rrs}}, \code{\link{rrst}} and \code{\link{msden}}.
}
\author{
T.M. Davies
}
