% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.density.R
\name{bivariate.density}
\alias{bivariate.density}
\alias{bivden}
\title{Bivariate kernel density/intensity estimation}
\usage{
bivariate.density(pp, h0, hp = NULL, adapt = FALSE, resolution = 128,
  gamma.scale = "geometric", edge = c("uniform", "diggle", "none"),
  intensity = FALSE, trim = 5, xy = NULL, pilot.density = NULL,
  leaveoneout = FALSE, parallelise = NULL, davies.baddeley = NULL,
  verbose = TRUE)
}
\arguments{
\item{pp}{An object of class \code{\link[spatstat]{ppp}} giving the observed
2D data set to be smoothed.}

\item{h0}{Global bandwidth for adaptive smoothing or fixed bandwidth for
constant smoothing. A numeric value > 0.}

\item{hp}{Pilot bandwidth (scalar, numeric > 0) to be used for fixed
bandwidth estimation of a pilot density in the case of adaptive smoothing.
If \code{NULL} (default), it will take on the value of \code{h0}. Ignored
when \code{adapt = FALSE} or if \code{pilot.density} is supplied as a
pre-defined pixel image.}

\item{adapt}{Logical value indicating whether to perform adaptive kernel
estimation. See `Details'.}

\item{resolution}{Numeric value > 0. Resolution of evaluation grid; the
density/intensity will be returned on a [\code{resolution} \eqn{\times}{x}
\code{resolution}] grid.}

\item{gamma.scale}{Scalar, numeric value > 0; controls rescaling of the
variable bandwidths. Defaults to the geometric mean of the bandwidth factors
given the pilot density (as per Silverman, 1986). See `Details'.}

\item{edge}{Character string giving the type of edge correction to employ.
\code{"uniform"} (default) corrects based on evaluation grid coordinate and
\code{"diggle"} reweights each observation-specific kernel. Setting
\code{edge = "none"} requests no edge correction. Further details can be
found in the documentation for \code{\link[spatstat]{density.ppp}}.}

\item{intensity}{Logical value indicating whether to return an intensity
estimate (integrates to the sample size over the study region), or a density
estimate (default, integrates to 1).}

\item{trim}{Numeric value > 0; controls bandwidth truncation for adaptive
estimation. See `Details'.}

\item{xy}{Optional alternative specification of the evaluation grid; matches
the argument of the same tag in \code{\link[spatstat]{as.mask}}. If
supplied, \code{resolution} is ignored.}

\item{pilot.density}{An optional pixel image (class
\code{\link[spatstat]{im}}) giving the pilot density to be used for
calculation of the variable bandwidths in adaptive estimation, \bold{or} a
\code{\link[spatstat]{ppp.object}} giving the data upon which to base a
fixed-bandwidth pilot estimate using \code{hp[1]}. If used, the pixel image
\emph{must} be defined over the same domain as the data given
\code{resolution} or the supplied pre-set \code{xy} evaluation grid;
\bold{or} the planar point pattern data must be defined with respect to the
same polygonal study region as in \code{pp}.}

\item{leaveoneout}{Logical value indicating whether to compute and return
the value of the density/intensity at each data point for an adaptive
estimate. See `Details'.}

\item{parallelise}{Numeric argument to invoke parallel processing, giving
the number of CPU cores to use when \code{leaveoneout = TRUE}. Experimental.
Test your system first using \code{parallel::detectCores()} to identify the
number of cores available to you.}

\item{davies.baddeley}{An optional numeric vector of length 3 to control
bandwidth partitioning for approximate adaptive estimation, giving the
quantile step values for the variable bandwidths for density/intensity and
edge correction surfaces and the resolution of the edge correction surface.
May also be provided as a single numeric value. See `Details'.}

\item{verbose}{Logical value indicating whether to print a function progress
bar to the console when \code{adapt = TRUE}.}
}
\value{
If \code{leaveoneout = FALSE}, an object of class \code{"bivden"}.
This is effectively a list with the following components: \item{z}{The
resulting density/intensity estimate, a pixel image object of class
\code{\link[spatstat]{im}}.} \item{h0}{A copy of the value of \code{h0}
used.} \item{hp}{A copy of the value of \code{hp} used.} \item{h}{A numeric
vector of length equal to the number of data points, giving the bandwidth
used for the corresponding observation in \code{pp}.} \item{him}{A pixel
image (class \code{\link[spatstat]{im}}), giving the `hypothetical' Abramson
bandwidth at each pixel coordinate conditional upon the observed data.
\code{NULL} for fixed-bandwidth estimates.} \item{q}{Edge-correction
weights; a pixel \code{\link[spatstat]{im}}age if \code{edge = "uniform"}, a
numeric vector if \code{edge = "diggle"}, and \code{NULL} if \code{edge =
"none"}.} \item{gamma}{The value of \eqn{\gamma} used in scaling the
bandwidths. \code{NA} if a fixed bandwidth estimate is computed.}
\item{geometric}{The geometric mean \eqn{G} of the untrimmed bandwidth
factors \eqn{\tilde{f}(x_i)^{-1/2}}. \code{NA} if a fixed bandwidth estimate
is computed.} \item{pp}{A copy of the \code{\link[spatstat]{ppp.object}}
initially passed to the \code{pp} argument, containing the data that were
smoothed.} \item{fromms}{A logical value indicating whether the object has
arisen from a call to \code{\link{multiscale.slice}}.} Else, if
\code{leaveoneout = TRUE}, simply a numeric vector of length equal to the
number of data points, giving the leave-one-out value of the function at the
corresponding coordinate.
}
\description{
Provides an isotropic adaptive or fixed bandwidth kernel density/intensity
estimate of bivariate/planar/2D data.
}
\examples{


\dontrun{
# to be filled
}

}
\references{
Abramson, I. (1982). On bandwidth variation in kernel estimates
--- a square root law, \emph{Annals of Statistics}, \bold{10}(4),
1217-1223.\cr\cr Davies, T.M. and Baddeley A. (2017), Fast computation of
spatially adaptive kernel estimates, \emph{Submitted}.\cr\cr Davies, T.M.
and Hazelton, M.L. (2010), Adaptive kernel estimation of spatial relative
risk, \emph{Statistics in Medicine}, \bold{29}(23) 2423-2437.\cr\cr Davies,
T.M., Jones, K. and Hazelton, M.L. (2016), Symmetric adaptive smoothing
regimens for estimation of the spatial relative risk function,
\emph{Computational Statistics & Data Analysis}, \bold{101}, 12-28.\cr\cr
Diggle, P.J. (1985), A kernel method for smoothing point process data,
\emph{Journal of the Royal Statistical Society, Series C}, \bold{34}(2),
138-147.\cr\cr Hall P. and Marron J.S. (1988) Variable window width kernel
density estimates of probability densities. \emph{Probability Theory and
Related Fields}, \bold{80}, 37-49.\cr\cr Marshall, J.C. and Hazelton, M.L.
(2010) Boundary kernels for adaptive density estimators on regions with
irregular boundaries, \emph{Journal of Multivariate Analysis}, \bold{101},
949-963.\cr\cr Silverman, B.W. (1986), \emph{Density Estimation for
Statistics and Data Analysis}, Chapman & Hall, New York. \cr\cr Wand, M.P.
and Jones, C.M., 1995. \emph{Kernel Smoothing}, Chapman & Hall, London.
}
\author{
T.M. Davies
}
