% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bivden.R
\name{print.bivden}
\alias{print.bivden}
\alias{print.rrs}
\alias{print.msden}
\title{Printing a bivariate density/intensity, relative risk surface, or
multi-scale density/intensity}
\usage{
\method{print}{bivden}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{bivden}}, \code{\link{rrs}}, or
\code{\link{msden}}.}

\item{...}{Ignored.}
}
\description{
\code{print} methods for classes \code{"\link{bivden}"}, \code{"\link{rrs}"},
and \code{"\link{msden}"}.
}
\author{
T.M. Davies
}
