\name{sparr-internal}
\Rdversion{1.1}
\alias{sparr-internal}
\alias{adens}
\alias{BAM.single}
\alias{BAM.prep}
\alias{bivden.LOO}
\alias{checkdb}
\alias{checkedge}
\alias{checkit}
\alias{checkran}
\alias{edgeh}
\alias{gethcats}
\alias{identical_windows}
\alias{K0}
\alias{LSCV.density.spatial.single}
\alias{LSCV.risk.single}
\alias{posifybivden}
\alias{processgamma}
\alias{processnstar}
\alias{processscaler}
\alias{point_image_by_bw}
\alias{taperoff}
\alias{tol.asy.ada}
\alias{tol.asy.fix}
\alias{tol.mc.ada}
\alias{tol.mc.fix}
\title{
Internal \code{sparr} functions
}
\description{
Internal \code{\link{sparr}} functions, only meant to be accessed by the user directly in a life-threatening situation.
}
\section{Usage}{
\code{adens(x,bwim,bwpts,resolution,edge,diggle,weights,intensity,hstep,qstep,qres,verbose)}\cr
\code{BAM.single(h,edge,BP)}\cr
\code{BAM.prep(cases,controls,lambda,erode,res)}\cr
\code{bivden.LOO(pp,h0,hp,gamma.scale,trim,resolution,parallel)}\cr
\code{checkdb(db)}\cr
\code{checkedge(edge)}\cr
\code{checkit(h,str)}\cr
\code{checkran(ran,nm)}\cr
\code{edgeh(bwim,pres,tres,step,W,verbose=FALSE)}\cr
\code{gethcats(h,breaks=NULL,step=0.05)}\cr
\code{identical_windows(w1,w2)}\cr
\code{KO(xy)}\cr
\code{LSCV.density.spatial.single(h,pp,res,edge)}\cr
\code{LSCV.risk.single(h,cases,controls,res,edge,hazey)}\cr
\code{posifybivden(x,eps=.Machine$double.xmin)}\cr
\code{processgamma(gamma.scale,pds)}\cr
\code{processnstar(n,pp)}\cr
\code{processscaler(s,pp)}\cr
\code{point_image_by_bw(h,hc,points,weights,WM)}\cr
\code{taperoff(x,zeropoint=0,onepoint=1,type=c("smooth", "cosine"))}\cr
\code{tol.asy.ada(f,g,beta,verbose=FALSE)}\cr
\code{tol.asy.fix(f,g,pooled,verbose=FALSE)}\cr
\code{tol.mc.ada(rs,ITER,parallelise,verbose,...)}\cr
\code{tol.mc.fix(rs,ITER,parallelise,verbose,...)}\cr
}
\keyword{internal}

\author{T.M. Davies}