% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bivden.R
\name{summary.bivden}
\alias{summary.bivden}
\alias{print.bivden}
\title{Summarising an estimated bivariate density/intensity object}
\usage{
\method{summary}{bivden}(object, ...)
}
\arguments{
\item{...}{Ignored.}

\item{x, object}{An object of class \code{\link{bivden}} resulting from a
call to \code{\link{bivariate.density}}.}
}
\description{
\code{print} and \code{summary} methods for class \code{\link{bivden}}.
}
\author{
T.M. Davies
}
