% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bivden.R
\name{summary.bivden}
\alias{summary.bivden}
\alias{summary.rrs}
\alias{summary.msden}
\title{Summarising a bivariate density/intensity, relative risk surface, or
multi-scale density/intensity}
\usage{
\method{summary}{bivden}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{bivden}}, \code{\link{rrs}}, or
\code{\link{msden}}.}

\item{...}{Ignored.}
}
\description{
\code{summary} methods for classes \code{"\link{bivden}"}, \code{"\link{rrs}"},
and \code{"\link{msden}"}.
}
\author{
T.M. Davies
}
