\name{summary.bivden}
\Rdversion{1.1}
\alias{summary.bivden}
\alias{print.bivden}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarising an estimated bivariate density/intensity object
}
\description{
\code{print} and \code{summary} methods for class \code{\link{bivden}}.
}
\usage{
\method{print}{bivden}(x, \dots)
\method{summary}{bivden}(object, \dots)
}
\arguments{
  \item{x, object}{
An object of class \code{\link{bivden}} resulting from a call to \code{\link{bivariate.density}}.
}
  \item{...}{
Ignored.
}
}

  
%\details{

%}
%\value{

%}
%\references{
%}
\author{
T.M. Davies
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%\code{\link{bivariate.density}}
%}
%\examples{
%KBivN(c(0.1,0.4))
%}

