\name{summary.msden}
\Rdversion{1.1}
\alias{summary.msden}
\alias{print.msden}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarising an estimated multi-scale density object
}
\description{
\code{print} and \code{summary} methods for class \code{\link{msden}}
}
\usage{
\method{print}{msden}(x, \dots)
\method{summary}{msden}(object, \dots)
}
\arguments{
  \item{x, object}{
An object of class \code{\link{msden}} resulting from a call to \code{\link{multiscale.density}}.
}
  \item{...}{
Ignored.
}
}

  
%\details{

%}
%\value{

%}
%\references{
%}
\author{
T.M. Davies
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%\code{\link{bivariate.density}}
%}
%\examples{
%KBivN(c(0.1,0.4))
%}

