\name{summary.rrs}
\Rdversion{1.1}
\alias{summary.rrs}
\alias{print.rrs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summarising an estimated relative risk function object
}
\description{
\code{print} and \code{summary} methods for class \code{\link{rrs}}.
}
\usage{
\method{print}{rrs}(x, \dots)
\method{summary}{rrs}(object, \dots)
}
\arguments{
  \item{x, object}{
An object of class \code{\link{rrs}} resulting from a call to \code{\link{risk}}.
}
  \item{...}{
Ignored.
}
}

  
%\details{

%}
%\value{

%}
%\references{
%}
\author{
T.M. Davies
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%\code{\link{bivariate.density}}
%}
%\examples{
%KBivN(c(0.1,0.4))
%}

